+++ src/main/java/gtclassic/common/tile/multi/GTTileMultiLESU.java
@@ -8,6 +8,7 @@ import gtclassic.common.GTLang;
 import gtclassic.common.container.GTContainerLESU;
 import gtclassic.common.util.GTIBlockFilters;
 import ic2.api.energy.tile.IMultiEnergySource;
+import ic2.core.IC2;
 import ic2.core.RotationList;
 import ic2.core.block.base.tile.TileEntityElectricBlock;
 import ic2.core.inventory.container.ContainerIC2;
@@ -15,12 +16,16 @@ import ic2.core.platform.lang.components.base.LocaleComp;
 import ic2.core.util.helpers.AabbUtil;
 import ic2.core.util.helpers.AabbUtil.BoundingBox;
 import ic2.core.util.helpers.AabbUtil.Processor;
+import ic2.core.util.obj.IClickable;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.init.SoundEvents;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.EnumFacing;
+import net.minecraft.util.EnumHand;
+import net.minecraftforge.fml.relauncher.Side;
 
 public class GTTileMultiLESU extends TileEntityElectricBlock
-		implements IMultiEnergySource, IGTMultiTileStatus, IGTDebuggableTile {
+		implements IMultiEnergySource, IGTMultiTileStatus, IClickable, IGTDebuggableTile {
 
 	private int blockCount;
 	public boolean enabled = true;
@@ -140,6 +145,34 @@ public class GTTileMultiLESU extends TileEntityElectricBlock
 	@Override
 	public boolean getStructureValid() {
 		return this.blockCount > 0;
+	}
+
+	@Override
+	public boolean hasRightClick() {
+		return true;
+	}
+
+	@Override
+	public boolean onRightClick(EntityPlayer player, EnumHand hand, EnumFacing facing, Side side) {
+		if (player.isSneaking() && player.getHeldItemMainhand().isEmpty()) {
+			this.onNetworkEvent(player, 0);
+			if (this.isSimulating()) {
+				IC2.platform.messagePlayer(player, this.getRedstoneMode());
+			}
+			player.playSound(SoundEvents.UI_BUTTON_CLICK, 1.0F, 1.0F);
+			return true;
+		}
+		return false;
+	}
+
+	@Override
+	public boolean hasLeftClick() {
+		return false;
+	}
+
+	@Override
+	public void onLeftClick(EntityPlayer var1, Side var2) {
+		// Needed for interface
 	}
 
 	@Override