+++ CHANGELOG.md
@@ -4,11 +4,15 @@
 
 # 1.1.4 (Not Released)
 - Added a very basic Nuclear Control style Display Screen, just single block form. Use a sensor stick to set block to read, will automatically parse useful information.
+- Note the Display Screens are very basic, and I have not decided how I want to handle lots of things like modular screens or adding information/titles yourself.   
 - Fixed a bug where relogging would cause the Basic Worktable to be stuck "in use" texture.
 - Changed Player Detector and Drum toggle modes to be sneaking with an empty hand instead of with a Wrench.
 - Added ability for IDSU, AESU, and LESU to output redstone level like IC2C EU storage tiles, sneak click with an empty hand to cycle modes.
 - Added comparator output for the Drum and the Battery Block.
 - Changed, you can now click the Matter Fabricator with a fluid container to empty it into the internal tank.
+- Changed, you can now click the Magic Energy Converter with a fluid container to empty it into the internal tank.
+- Fixed crash with AE2 particle effect ASM
+- Fixed Tesseract Terminal updating fluid capability but not items, now updates item pipes and stuff if destination is invalid.
 
 # 1.1.3
 - Changed, Magical Energy Absorber can now siphon energy from End Portals with a few very specific behaviors. only one per portal or KABOOM! (configurable). Also if you absorb portals in Overworld Strongholds there is a chance to disable the portal and remove an Ender Eye from the portal frame every so often (roughly one MC day).